Add restriction on enrollment start_date and end_date
The enrollment start_date and end_date will have default value as
the startDate and endDate in the selected program.

If the program has no or empty endDate, the program is considered
to be never ending.

Also, the enrollment start_date must be later/equal than the
program start_date, and the enrollment end_date must be eariler or
equal to the program end_date

And the program-id in the program filter is deleted, now showing
only program name