Add enrollment overriding fee in storage and display
Note that any future enrollment will always has the billing_fee.
If the provider did not change during enrolling, it will be the
default of the program fee.
If the provider did change it, it will be the overriding value.

Overriding will be allowed after the invoice calculation also has
updated logic.